Make TestRunner a single type
